breaking-api: Rename InputField to NumberInputField
- Because it isn't a generic text field, so it's much more descriptive this way
